Refactored the code that creates BSP connections.
